Add inital support for unsigned types

+++ tests/test1/test1.c
@@ -2,6 +2,6 @@
 int main() {
     long i = 1;
     int j = 2;
-    int k = 3;
+    unsigned long long k = 3;
     return i + j - k;
 }
